requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
Pillow==9.5.0
django-storages==1.14.0
boto3==1.26.0
razorpay==1.3.0
celery==5.3.0
redis==4.5.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
gunicorn==20.1.0
whitenoise==6.4.0
django-ratelimit==4.0.0
python-json-logger==2.0.0
sentry-sdk==1.27.0
stripe==5.4.0
django-filter==23.1
djangorestframework-simplejwt==5.2.2
pytz==2023.3
requests==2.31.0
